refactor: change kafka analytics to mesure increment between two reports
instead of use TSDS with gauge and counter
BREAKING CHANGE: change kafka analytics to mesure increment between two reports
